изменил систему расчёта денег

+++ app/src/main/java/ru/sibur/android/garbagecollector/Automata.java
@@ -5,7 +5,7 @@ package ru.sibur.android.garbagecollector;
  */
 
 public class Automata extends ShopItem {
-    Automata(String nameIN, float priceIn){
+    Automata(String nameIN, int priceIn){
         super(nameIN, priceIn);
     }
 }
