+++ app/src/main/java/ru/sibur/android/garbagecollector/AutomataCostReduce.java
@@ -3,7 +3,7 @@ package ru.sibur.android.garbagecollector;
 import ru.sibur.android.garbagecollector.Upgrade;
 
 public class AutomataCostReduce extends Upgrade {
-    AutomataCostReduce(String nameIn, float priceIn){
+    AutomataCostReduce(String nameIn, int priceIn){
         super(nameIn, priceIn);
     }
 }
