+++ app/src/main/java/ru/sibur/android/garbagecollector/AutomataPerfomanceUpgrade.java
@@ -4,7 +4,7 @@ package ru.sibur.android.garbagecollector;
 import ru.sibur.android.garbagecollector.Upgrade;
 
 public class AutomataPerfomanceUpgrade extends Upgrade {
-    AutomataPerfomanceUpgrade(String nameIn, float priceIn){
+    AutomataPerfomanceUpgrade(String nameIn, int priceIn){
         super(nameIn, priceIn);
     }
 }
